Add trading module
Trading is now possible between planets
In the map screen, a planet can be jumped to. "T" can then be
pressed to bring up the trading screen. The arrows are used to
move the selector and either buy or sell one unit of commodity
at a time. "Q" can be pressed to trade as much as possible.
For now, jumping between planets does not reduce range

Signed-off-by: TarrielDRC <tarriel345@gmail.com>